feat: add user roles
Add the 'role' column to the users table.
Add the 'role' field to the User model's fillable array.
Automatically cast the User model's 'role' value to an instance of the
UserRole Enum.
Create the UserRole Enum to specify acceptable user role values.
Update the User model factory to allow for the creation of users with
either a 'customer' or 'admin' role.